# Remove use of a global lock