--- a/context.go
+++ b/context.go
@@ -5,87 +5,94 @@
 package context
 
 import (
+	"io"
 	"net/http"
-	"sync"
-	"time"
 )
 
-var (
-	mutex sync.RWMutex
-	data  = make(map[*http.Request]map[interface{}]interface{})
-	datat = make(map[*http.Request]int64)
-)
+type attributeStore map[interface{}]interface{}
+
+type decoratedReadCloser struct {
+	body       io.ReadCloser
+	attributes *attributeStore
+}
+
+func decorate(r *http.Request) *decoratedReadCloser {
+	result := decoratedReadCloser{body: r.Body}
+	r.Body = &result
+	return &result
+}
+
+func (w *decoratedReadCloser) Read(p []byte) (n int, err error) {
+	return w.body.Read(p)
+}
+
+func (w *decoratedReadCloser) Close() error {
+	return w.body.Close()
+}
+
+func attributes(r *http.Request) attributeStore {
+	if atts, ok := peekAttributes(r); ok {
+		return *atts
+	}
+
+	atts := make(attributeStore)
+
+	// decorate the request
+	decorator := decorate(r)
+	decorator.attributes = &atts
+
+	return atts
+}
+
+func peekAttributes(r *http.Request) (*attributeStore, bool) {
+	if v, ok := r.Body.(*decoratedReadCloser); ok {
+		return v.attributes, true
+	}
+	return nil, false
+}
 
 // Set stores a value for a given key in a given request.
 func Set(r *http.Request, key, val interface{}) {
-	mutex.Lock()
-	if data[r] == nil {
-		data[r] = make(map[interface{}]interface{})
-		datat[r] = time.Now().Unix()
-	}
-	data[r][key] = val
-	mutex.Unlock()
+	data := attributes(r)
+	data[key] = val
 }
 
 // Get returns a value stored for a given key in a given request.
 func Get(r *http.Request, key interface{}) interface{} {
-	mutex.RLock()
-	if ctx := data[r]; ctx != nil {
-		value := ctx[key]
-		mutex.RUnlock()
-		return value
+	data := attributes(r)
+	if v, ok := data[key]; ok {
+		return v
 	}
-	mutex.RUnlock()
 	return nil
 }
 
 // GetOk returns stored value and presence state like multi-value return of map access.
 func GetOk(r *http.Request, key interface{}) (interface{}, bool) {
-	mutex.RLock()
-	if _, ok := data[r]; ok {
-		value, ok := data[r][key]
-		mutex.RUnlock()
-		return value, ok
-	}
-	mutex.RUnlock()
-	return nil, false
+	data := attributes(r)
+	v, ok := data[key]
+	return v, ok
 }
 
 // GetAll returns all stored values for the request as a map. Nil is returned for invalid requests.
 func GetAll(r *http.Request) map[interface{}]interface{} {
-	mutex.RLock()
-	if context, ok := data[r]; ok {
-		result := make(map[interface{}]interface{}, len(context))
-		for k, v := range context {
-			result[k] = v
-		}
-		mutex.RUnlock()
-		return result
+	if atts, ok := peekAttributes(r); ok {
+		return *atts
 	}
-	mutex.RUnlock()
 	return nil
 }
 
 // GetAllOk returns all stored values for the request as a map and a boolean value that indicates if
 // the request was registered.
 func GetAllOk(r *http.Request) (map[interface{}]interface{}, bool) {
-	mutex.RLock()
-	context, ok := data[r]
-	result := make(map[interface{}]interface{}, len(context))
-	for k, v := range context {
-		result[k] = v
+	if atts, ok := peekAttributes(r); ok {
+		return *atts, ok
 	}
-	mutex.RUnlock()
-	return result, ok
+	return nil, false
 }
 
 // Delete removes a value stored for a given key in a given request.
 func Delete(r *http.Request, key interface{}) {
-	mutex.Lock()
-	if data[r] != nil {
-		delete(data[r], key)
-	}
-	mutex.Unlock()
+	delete(attributes(r), key)
 }
 
 // Clear removes all values stored for a given request.
@@ -93,15 +100,15 @@ func Delete(r *http.Request, key interface{}) {
 // This is usually called by a handler wrapper to clean up request
 // variables at the end of a request lifetime. See ClearHandler().
 func Clear(r *http.Request) {
-	mutex.Lock()
 	clear(r)
-	mutex.Unlock()
 }
 
 // clear is Clear without the lock.
 func clear(r *http.Request) {
-	delete(data, r)
-	delete(datat, r)
+	data := attributes(r)
+	for k := range data {
+		delete(data, k)
+	}
 }
 
 // Purge removes request data stored for longer than maxAge, in seconds.
@@ -114,23 +121,7 @@ func clear(r *http.Request) {
 // amount of memory. In case this is detected, Purge() must be called
 // periodically until the problem is fixed.
 func Purge(maxAge int) int {
-	mutex.Lock()
-	count := 0
-	if maxAge <= 0 {
-		count = len(data)
-		data = make(map[*http.Request]map[interface{}]interface{})
-		datat = make(map[*http.Request]int64)
-	} else {
-		min := time.Now().Unix() - int64(maxAge)
-		for r := range data {
-			if datat[r] < min {
-				clear(r)
-				count++
-			}
-		}
-	}
-	mutex.Unlock()
-	return count
+	return 0
 }
 
 // ClearHandler wraps an http.Handler and clears request values at the end