Request aren’t pooled, so we can wrap the request Body and store
attributes there. This means that each request has its own place to
store attributes, without having to go through a mutex. It also means
that Clear can be dropped in a later release, since the Request will be
eligible for GC and should do our job for us.


Before:
```
> go test -bench .
PASS
BenchmarkMutexSameReadWrite1	  100000	     13965 ns/op
BenchmarkMutexSameReadWrite2	   50000	     28821 ns/op
BenchmarkMutexSameReadWrite4	   30000	     57187 ns/op
BenchmarkMutex1	   20000	     86459 ns/op
BenchmarkMutex2	   10000	    222826 ns/op
BenchmarkMutex3	   20000	     90147 ns/op
BenchmarkMutex4	     200	   7688993 ns/op
BenchmarkMutex5	      30	  50292186 ns/op
BenchmarkMutex6	       5	 298536790 ns/op
ok  	_/Users/jabley/git/context	19.730s
```

With the mutex removed:
```
> go test -bench .
PASS
BenchmarkMutexSameReadWrite1	  100000	     12700 ns/op
BenchmarkMutexSameReadWrite2	   50000	     25692 ns/op
BenchmarkMutexSameReadWrite4	   30000	     50888 ns/op
BenchmarkMutex1	   20000	     71686 ns/op
BenchmarkMutex2	   10000	    197388 ns/op
BenchmarkMutex3	   20000	     77450 ns/op
BenchmarkMutex4	     200	   6996854 ns/op
BenchmarkMutex5	      30	  42156099 ns/op
BenchmarkMutex6	       5	 261614479 ns/op
ok  	_/Users/jabley/git/context	17.295s
```