@kisielk 2015-01-15
If going with this scheme, I think the attributes on the decorator need a mutex. It's possible the request will get passed to multiple goroutines an they will be concurrently accessing the context.