@win-t 2015-07-18
you say It's possible the request will get passed to multiple goroutines, but why do that?,
Isn't that one connection supposed to be handled by one goroutine?