@kisielk 2015-07-18
Definitely, but currently there is no such restriction so it's possible that people are passing it to multiple goroutines. If we change the behaviour we are introducing race conditions in to code which does that.