@kisielk 2015-08-10
Okay, I had another look at this and I don't think the concurrency is really an issue. The Request type doesn't protect anything, so it can be assumed that writing to it concurrently is bound to introduce races. One thing I am concerned about though is that it seems that internally the net/http server does some wrapping of the request Body in its own types and then does checks against them later. For example there's a special type `body` internal to the package which supports early close of connections, and another one called `expectContinueReader` ... I'm not sure if this change could possibly interfere with those?