@elithrar 2015-08-11
Okay, took a look at this: the performance increase is reasonable. Roughly ~7% (63k req/s vs 69 req/s @ GOMAXPROCS=4) in the absolute worst case for the mutex-protected context (aka real world will be less than that). The benchmark just sets/gets a small struct out of the context on every request and does no other meaningful work, so it's thrashing the lock pretty hard.

I also ran a super simple test to see what kind of issues we would expect (i.e. the ones raised by @kisielk) with this approach and other libs that wrap or copy the request.

```go
package main

import (
	"fmt"
	"net/http"
	"net/http/httputil"

	"github.com/gorilla/context"
	"github.com/gorilla/mux"
)

const userKey string = "attest.user"

type User struct {
	ID    string
	Email string
}

func SomeHandler(w http.ResponseWriter, r *http.Request) {
	b, err := httputil.DumpRequest(r, true)
	if err != nil {
		http.Error(w, err.Error(), 500)
		return
	}

	val, ok := context.GetOk(r, userKey)
	if !ok {
		http.Error(w, "No key in map", 500)
		return
	}

	user, ok := val.(*User)
	if !ok {
		http.Error(w, "No user", 500)
		return
	}

	fmt.Fprintf(w, "req: %s\nx user: %v\n", &b, user)
}

func SomeMiddleware(h http.Handler) http.Handler {
	fn := func(w http.ResponseWriter, r *http.Request) {
		user := &User{"1831831831", "matt@eatsleeprepeat.net"}
		context.Set(r, userKey, user)
		h.ServeHTTP(w, r)
	}

	return http.HandlerFunc(fn)
}

func main() {
	r := mux.NewRouter()
	r.HandleFunc("/", SomeHandler)
	http.ListenAndServe(":8000",
		SomeMiddleware(r))
}
```

* Running this with the current master (mutex and all) results in a dumped response and a dumped user object - the correct behaviour:

```sh

~/.go/src/github.com/gorilla/context(master ✔) curl http://localhost:8000/
req: &GET / HTTP/1.1
Host: localhost:8000
Accept: */*
User-Agent: curl/7.37.1


user: &{1831831831 matt@eatsleeprepeat.net}
```

* Running this against the `feature/remove-global-lock` branch:

```sh
~/.go/src/github.com/gorilla/context(feature/remove-global-lock ✔) curl http://localhost:8000/
No key in map
```

Given the above I'd want to do more digging before we commit this. Any middleware that copies `r.Body` doesn't get the context map with that copy.