@goodfoo 2016-02-26
@jabley Hi, I used your idea to embed golang.org/x/net/context in the request
https://github.com/goodfoo/requestcontext
Thank you