@elithrar 2016-02-26
I'm going to close this out as:

1. The performance benefits are minor, and gorilla/context is rarely the bottleneck in a web service vs. DB calls, network calls, JSON rendering.
2. It conflicts with other libs that decorate the request inc. parts of the std. lib.

Happy to re-open if there's a solution to no. 2.